+++ src/main/java/com/rlti/autoescola/matricula/application/api/MatriculaListResponse.java
@@ -14,7 +14,7 @@ public class MatriculaListResponse {
     UUID idMatricula;
     String cpf;
     TipoPagamento tipoPagamento;
-    Double valorEntrada;
+    BigDecimal valorEntrada;
     int desconto;
     int quantidadeParcelas;
     BigDecimal valorFinal;
